flask_requirements.txt
# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Core ML Libraries
torch==2.1.0
torchaudio==2.1.0
transformers==4.36.0

# Speech Processing
openai-whisper==20231117
librosa==0.10.1
sounddevice==0.4.6
soundfile==0.12.1

# NLP & Text Processing
nltk==3.8.1
wordcloud==1.9.3

# Translation & TTS
gtts==2.5.0

# Scientific Computing
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2

# PDF Generation
reportlab==4.0.7

# Data Handling
pandas==2.1.4

# Additional Dependencies
Pillow==10.1.0
python-dateutil==2.8.2
